rename `gen` to `generator` for edition 2025 compatibility
