+++ examples/secure_request_guard/src/oauth2.rs
@@ -10,7 +10,7 @@ use rocket_okapi::okapi::openapi3::{
     OAuthFlows, Object, SecurityRequirement, SecurityScheme, SecuritySchemeData,
 };
 use rocket_okapi::{
-    gen::OpenApiGenerator,
+    generator::OpenApiGenerator,
     openapi,
     request::{OpenApiFromRequest, RequestHeaderInput},
 };
