+++ examples/secure_request_guard/src/open_id.rs
@@ -9,7 +9,7 @@ use rocket_okapi::okapi::openapi3::{
     Object, SecurityRequirement, SecurityScheme, SecuritySchemeData,
 };
 use rocket_okapi::{
-    gen::OpenApiGenerator,
+    generator::OpenApiGenerator,
     openapi,
     request::{OpenApiFromRequest, RequestHeaderInput},
 };
