+++ rocket-okapi/CHANGELOG.md
@@ -7,6 +7,7 @@ This project follows the [Semantic Versioning standard](https://semver.org/).
 ### Added
 
 ### Changed
+- Renamed `gen` module to `generator` for Edition 2025 compatibility
 
 ### Deprecated
