+++ rocket-okapi/src/request/from_segments_impls.rs
@@ -2,7 +2,7 @@
 //! <https://docs.rs/rocket/latest/rocket/request/trait.FromSegments.html#foreign-impls>
 
 use super::OpenApiFromSegments;
-use crate::gen::OpenApiGenerator;
+use crate::generator::OpenApiGenerator;
 use okapi::openapi3::{Object, Parameter, ParameterValue};
 use schemars::JsonSchema;
