+++ rocket-okapi/src/response/mod.rs
@@ -1,6 +1,6 @@
 mod responder_impls;
 
-use super::gen::OpenApiGenerator;
+use super::generator::OpenApiGenerator;
 use super::Result;
 use okapi::openapi3::Responses;
